requirements.txt
azure_storage==0.37.0
fastapi==0.115.5
joblib==1.2.0
langchain==0.3.7
lightgbm==4.5.0
matplotlib==3.7.2
mlflow==2.17.2
mlflow_skinny==2.17.2
numpy==1.24.3
pandas==2.0.3
prophet==1.1.6
pydantic==2.9.2
PyYAML==6.0
PyYAML==6.0.2
scikit_learn==1.3.0
scipy==1.14.1
statsmodels==0.14.0
yfinance==0.2.49
